remove old queue structure - migrate to sqlite3 db:
