+++ auto-kol/src/index.ts
@@ -1,8 +1,7 @@
 import express from 'express';
 import { config } from './config/index.js';
 import { createLogger } from './utils/logger.js';
-import { getAllSkippedTweetsMemory, getSkippedTweetMemory, moveToQueueMemory } from './services/queue/index.js';
-import { updateResponseStatus, getAllPendingResponses } from './services/database/index.js';
+import { updateResponseStatus, getAllPendingResponses, moveSkippedToQueue } from './services/database/index.js';
 import { runWorkflow } from './services/agents/workflow.js';
 import { twitterClientScraper } from './services/twitter/apiv2.js';   
 import { initializeSchema, initializeDefaultKOLs, initializeDatabase, addDsn } from './database/index.js';
@@ -154,52 +153,52 @@ const startServer = () => {
     });
 
     // Get all skipped tweets
-    app.get('/tweets/skipped', (_, res) => {
-        const skippedTweets = getAllSkippedTweetsMemory();
-        res.json(skippedTweets);
-    });
+    // app.get('/tweets/skipped', (_, res) => {
+    //     const skippedTweets = getAllSkippedTweetsMemory();
+    //     res.json(skippedTweets);
+    // });
 
     // Get specific skipped tweet
-    app.get('/tweets/skipped/:id', (req, res) => {
-        const skipped = getSkippedTweetMemory(req.params.id);
-        if (!skipped) {
-            return res.status(404).json({ error: 'Skipped tweet not found' });
-        }
-        res.json(skipped);
-    });
+    // app.get('/tweets/skipped/:id', (req, res) => {
+    //     const skipped = getSkippedTweetMemory(req.params.id);
+    //     if (!skipped) {
+    //         return res.status(404).json({ error: 'Skipped tweet not found' });
+    //     }
+    //     res.json(skipped);
+    // });
 
-    // Move skipped tweet to response queue
-    app.post('/tweets/skipped/:id/queue', async (req, res) => {
-        try {
-            const skipped = getSkippedTweetMemory(req.params.id);
-            if (!skipped) {
-                return res.status(404).json({ error: 'Skipped tweet not found' });
-            }
+    // // Move skipped tweet to response queue
+    // app.post('/tweets/skipped/:id/queue', async (req, res) => {
+    //     try {
+    //         const skipped = getSkippedTweetMemory(req.params.id);
+    //         if (!skipped) {
+    //             return res.status(404).json({ error: 'Skipped tweet not found' });
+    //         }
 
-            const { response } = req.body;
-            if (!response) {
-                return res.status(400).json({ error: 'Response is required' });
-            }
+    //         const { response } = req.body;
+    //         if (!response) {
+    //             return res.status(400).json({ error: 'Response is required' });
+    //         }
 
-            const queuedResponse = await moveToQueueMemory(skipped.id, {
-                id: skipped.id,
-                tweet: skipped.tweet,
-                response: {
-                    content: response.content,
-                    references: response.references
-                },
-                workflowState: skipped.workflowState,
-                created_at: new Date(),
-                updatedAt: new Date(),
-                status: 'pending'
-            });
+    //         const queuedResponse = await moveSkippedToQueue(skipped.id, {
+    //             id: skipped.id,
+    //             tweet: skipped.tweet,
+    //             response: {
+    //                 content: response.content,
+    //                 references: response.references
+    //             },
+    //             workflowState: skipped.workflowState,
+    //             created_at: new Date(),
+    //             updatedAt: new Date(),
+    //             status: 'pending'
+    //         });
 
-            res.json(queuedResponse);
-        } catch (error) {
-            logger.error('Error moving skipped tweet to queue:', error);
-            res.status(500).json({ error: 'Failed to move tweet to queue' });
-        }
-    });
+    //         res.json(queuedResponse);
+    //     } catch (error) {
+    //         logger.error('Error moving skipped tweet to queue:', error);
+    //         res.status(500).json({ error: 'Failed to move tweet to queue' });
+    //     }
+    // });
 
     // Start server
     app.listen(config.PORT, () => {
